Fixed tax not being set for newly added cart items
A newly added item in the cart does not have an ID before it/quote gets saved.
TaxJar API call was using $item->getId() to identify the items,
which failed in this case and caused no tax being set on the new item.

This fix uses the index in a shared array to identify the items instead.